+++ src/ts/markdown/chartRender.ts
@@ -11,11 +11,15 @@ export const chartRender = (element: (HTMLElement | Document) = document,
     if (echartsElements.length > 0) {
         addScript(`${cdn}/dist/js/echarts/echarts.min.js`, "vditorEchartsScript").then(() => {
             echartsElements.forEach((e: HTMLDivElement) => {
+                const text = e.innerText.trim();
+                if (!text) {
+                    return;
+                }
                 try {
                     if (e.getAttribute("data-processed") === "true") {
                         return;
                     }
-                    const option = JSON.parse(e.innerText.trim());
+                    const option = JSON.parse(text);
                     echarts.init(e).setOption(option);
                     e.setAttribute("data-processed", "true");
                 } catch (error) {